Create module for rendering App funcion
