+++ src/index.js
@@ -1,14 +1,15 @@
 /* eslint-disable no-restricted-globals */
 /* eslint-disable no-restricted-syntax */
-/* eslint-disable guard-for-in */
-/* eslint-disable radix */
+
 import './style.css';
-import * as module from './utils/module1.js';
+import * as Local from './utils/storage.js';
+import Render from './utils/rendering.js'
 import EditTask from './utils/editing.js';
 
-const bullets = module.Storage.getFromStorage();
+const bullets = Local.Storage.getFromStorage();
 
-window.addEventListener('load', module.renderList(bullets));
+window.addEventListener('load', Render.renderList(bullets));
+
 
 // EVENT LISTENERS
 
@@ -16,8 +17,8 @@ const addTask = document.querySelector('#addTask');
 addTask.addEventListener('submit', (e) => {
   e.preventDefault();
   EditTask.add(bullets);
-  module.Storage.saveToStorage(bullets);
-  module.renderList(bullets);
+  Local.Storage.saveToStorage(bullets);
+  Render.renderList(bullets);
   location.reload();
 });
 
@@ -27,10 +28,10 @@ taskList.addEventListener('change', (e) => {
   if (e.target.classList.contains('status')) {
     const { id } = e.target.parentElement;
     const taskBody = document.getElementById(`task-${id}`);
-    module.Status.toggleBullet(bullets, parseInt(id, 10));
-    module.Storage.saveToStorage(bullets);
+    Local.Status.toggleBullet(bullets, parseInt(id, 10));
+    // Local.Storage.saveToStorage(bullets);
     taskBody.classList.toggle('completed');
-    module.Storage.saveToStorage(bullets);
+    Local.Storage.saveToStorage(bullets);
   }
 });
 
@@ -41,7 +42,7 @@ inputs.forEach((input) => {
     const id = parseInt(e.target.parentElement.id);
     const { value } = e.target;
     EditTask.updateTask(bullets, id, value);
-    module.Storage.saveToStorage(bullets);
+    Local.Storage.saveToStorage(bullets);
   });
 });
 
@@ -72,8 +73,8 @@ function updateId() {
   bullets.forEach((task, index) => {
     task.id = index + 1;
   });
-  module.Storage.saveToStorage(bullets);
-  module.renderList(bullets);
+  Local.Storage.saveToStorage(bullets);
+  Render.renderList(bullets);
   location.reload();
 }
 
@@ -92,6 +93,6 @@ const clearBtn = document.querySelector('.clearBtn');
 clearBtn.addEventListener('click', () => {
   EditTask.clearCompleted(bullets);
   updateId();
-  module.Storage.saveToStorage(bullets);
-  module.renderList(bullets);
+  Local.Storage.saveToStorage(bullets);
+  Render.renderList(bullets);
 });
