+++ src/utils/rendering.js
@@ -1,31 +1,4 @@
-/* eslint-disable max-classes-per-file */
-
-class Storage {
-  static saveToStorage(arr) {
-    localStorage.setItem('todoList', JSON.stringify(arr));
-  }
-
-  static getFromStorage() {
-    let tasks;
-    if (localStorage.getItem('todoList') === null) {
-      tasks = [];
-    } else {
-      tasks = JSON.parse(localStorage.getItem('todoList'));
-    }
-    return tasks;
-  }
-}
-
-class Status {
-  static toggleBullet = (arr, id) => {
-    const bullet = arr.find((task) => task.id === id);
-    if (bullet) {
-      bullet.completed = !bullet.completed;
-    }
-  }
-}
-
-function renderList(arr) {
+export default function renderList(arr) {
   const list = document.getElementById('listContainer');
   list.innerHTML = '';
   if (arr.length === 0) {
@@ -55,5 +28,3 @@ function renderList(arr) {
     });
   }
 }
-
-export { Status, Storage, renderList };
